#64 Fix UpdateWaitList
As per requirements, we are to remove a user from a waitlist only when
that use obtained or made a reservation on the same timeslot (date and
hour) for a different room. Thus, the method should thus be taking date
and hour as parameter. Those updates were tested with many scenarios.